remove a potential race in turnitin processing; fixes #8704
Previously we spawned a new job to check the status of each asset sent to
turnitin.  If there were multiple files it was possible that two of the jobs
could run at the same time, with one overwriting the score computed by the
other.  This commit changes the logic so that only one job is spawned, which
checks the status on all outstanding assets.

test plan:
- turn in multiple files to a turnitin assignment
- ensure you get scores for both assignments

Change-Id: I78edb33584b8cf99d490681154b32559c0a7db25
Reviewed-on: https://gerrit.instructure.com/11722
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>