correctly associate group assignment submissions
when a student uploads a file as a submission to a group assignment,
associate the attachment with the group, not with the individual
student. this was happening correctly for attachments created through
the submissions_controller, but not for those where the attachment was
created through the files controller before being associated with the
submission in the submissions controller.

fixes #8965

test-plan:
 - ensure s3 uploads are enabled
 - create a course with at least two students (e.g. Jane and Bob)
 - create a group category in that course
 - create at least one group in that category
 - assign both Jane and Bob to the same group in that category
 - create a group assignment associated with that category with file
   uploads allowed
 - masquerade as Jane
 - go to the assignment and upload a file
 - stop masquerading
 - go to the speedgrader for the assignment
 - both Jane and Bob should show the uploaded submission

Change-Id: I4d837fe35fe318ea216e2d3af71d3332115a9ae3
Reviewed-on: https://gerrit.instructure.com/11840
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>