better styling for empty comments fields when submitting an assignment
The "Comments:" blabel was unnecessary and actually kind of confusing,
so I replaced with just a textarea that has a placeholder and a
title. The labels were also not accessible since all the comments
textareas had the same ids, so all but one label were referencing
the wrong textarea.

test plan:
- make sure you can still make comments when submitting an assignment

Change-Id: I9fc97d7827d564cb9cfb7c5903ac1e4301b8ef9c
Reviewed-on: https://gerrit.instructure.com/12040
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>