fix ordering of ui-dialog buttons and make primary ones blue again
our old jquery ui theme css was float:right'ing these
which means to have them show up in the right order
you would have to create them in the wrong order.
we fixed that with the new jquery-ui-theme but now
we need to fix places where they were in the wrong
order in code.

also the class name to make them blue has changed
so I fixed that too

Change-Id: If0cbebce4788b0c68cd71a08cd2880076e139bc0
Reviewed-on: https://gerrit.instructure.com/12214
Reviewed-by: Joe Tanner <joe@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>