fix uploading files on Chrome v21 on Windows
For some reason, uploadify on Chrome Windows as of version 21 appears to be
calling the uploadify error callback when the upload request actually
succeeded. This might be a bug in uploadify, but the upgrade to a more recent
uploadify is still a work in progress.

This intercepts the success in the failure callback and kicks off what would
normally happen on success.

fixes #9766

test plan:
 * Install Windows and Chrome v21
 * Enable S3
 * Upload a file in the Files section using "Add File" link
 * The upload should succeed

Change-Id: If254182dfc35c8a67deff8be5e7ebeff1c18ada3
Reviewed-on: https://gerrit.instructure.com/12744
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>