don't load all enrollments on xhr to course#show
fixes #10003

when grade publishing was disabled, a strange series of events was causing the
course settings page to make a GET xhr request to the course show action, which
handles these types of requests differently and returns the course json with
ALL current enrollments attached.

this sort of defeats the purpose of paginating the enrollments on the course
settings page, so that is sad.

test plan:
- with the network tab open, go to the course settings page in a course with
  lots of students
- you shouldn't see the offending xhr request
- grade publishing should still work

Change-Id: I3348d312ec6ad78b2a1ce9c1c2a9a9685f2f5a19
Reviewed-on: https://gerrit.instructure.com/12982
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>