make quiz show "correct answer" feedback as teacher reviews student submission. Fixes #10295
When a quiz is set to not show correct answers, a teacher, while reviewing
a student submission, should see the "Correct Answer" arrows and feedback.
However, when a teacher is reviewing their own submission or doing a preview,
it should respect the quiz setting and not display the correct answer feedback.

Also in this commit is a minor display fix to inputs where the width was set too
narrow on the results page.

Testing Notes:
==========
* As a teacher, setup a quiz to NOT show correct answers
  to students.
* As the teacher, take the quiz and verify that it this setting is respected
  when viewing the teacher's own submission or when previewing the quiz.
* As the teacher, review a student's submission and verify that the
  quiz displays the "correct answer" feedback even though the quiz
  is configured to not display it.
* Verify that the "correct answer" feedback shows up for a quiz
  that is configured to display it.

Change-Id: Ibc939918db22c3add8fc00f18a50de03123d811a
Reviewed-on: https://gerrit.instructure.com/13368
Reviewed-by: Zach Wily <zach@instructure.com>
Tested-by: Zach Wily <zach@instructure.com>