don't display "select correct answer" arrows for fill in multiple blanks. Fixes #10325
Hide the "select correct answer" arrows that display next to
answers but only for "Fill in multiple blanks" question types.
All answers listed are considered correct and it doesn't make
sense to allow selecting a "correct" one. Doing so would
cause the other answers to appear as incorrect (red) but
it would still not treat the answers as incorrect. This change
makes it more consistent with current "fill in the blank" question
types.

Testing Notes:
==========
* Create/edit a quiz
* Create/edit a "Fill in multiple blanks" question
* Verify that the arrows for selecting a correct
   answer no longer appear.
* Verify that the arrows appear for other questions
   that need ability to select correct answers.

Change-Id: Ieadd5667d5a349e9693771f43c01aac36f9327f6
Reviewed-on: https://gerrit.instructure.com/13541
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>