properly escape course name in course copy form
fixes #11147

test plan: Create a course with a name like
"><img src=x onerror=prompt(1);>
Then go to copy the course. In the "adjust dates" section of the form,
you should see that string as the course name, rather than have the
browser interpret it as HTML and execute the javascript snippet.

Change-Id: Ic632088c782d8707f970f70543c3e323a35b5b67
Reviewed-on: https://gerrit.instructure.com/14218
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>