stabilize paginated calendar event urls
both the server and the javascript were adding extra copies of the bulk of
the query parameters for each subsequent page of events. stop it. now,
the request for the second page should only differ from the first in the
presence of a page=2 parameter, and the remaining page requests should
differ from the second only in the value of the page parameter.

this was particularly bad, because for a student with even a minor
number of courses, it would quickly bloat the url length to where apache
would barf.

fixes #11023

test-plan:
 - create a user
 - create 51 assignments (or events, but not "assignments or events") in
   that user's calendar for the current month in order to trigger
   pagination
 - view the user's calendar2 page for the current month, watching the
   xhr requests generated
 - the url for the request for the second page of events should only
   differ from that of the first page by the page parameter (other
   parameters may end up reordered, but should retain presence and
   value)

Change-Id: Ia2fb7511db83eed964b1d05ca13e0ab49e71b9ac
Reviewed-on: https://gerrit.instructure.com/14289
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Willesen <jonw@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>