[#5773] Ruby 1.9.3 spec fixes
[#5773] correctly evaluate the custom_fields_string

My impression is that order is not enforced in the
hash being used to build the fields string, as 
this was failing in 1.8.7 for me (but not in 
1.9.3), because the string being checked was
"b=456\na=123" while the string being returned
from the custom_fields_string method was 
"a=123\nb=456".  Testing with a regex instead
confirms the content is there without relying on
any particular order. [EDIT: Per Brian P, decided
to sort the params instead to enforce order]

[#5773] fix admin_settings_tab_spec for ruby 1.9.3

String's "each" method was being used, which is 
gone in 1.9.  Fixed by performing a guard clause 
in the check_box_verifier in the spec file to 
conditionally take a string argument and convert 
it to an array.  Also converted the rest of the 
method to assume an array value for that argument.

[#5773] fix collaborations selenium spec for 1.9.3

This is an instance of using the "each" method 
of the string class which is no longer present in 
1.9 in one of the helper methods in this spec.  
Erected a guard clause to check for Array-ness 
and to replace the argument with an 
array if necessary.

[#5773] fix courses show view

There was a spot in this view where it was 
interpolating "self.id" into an html attribute
rather than "@instance_model.id".  This sort
of worked in ruby 1.8, where it would interpolate
the object id of the view instance.  In 1.9,
that id becomes "object_id", which is what
one would have to change the code to in order
to preserve the current behavior.  
However, it seems more correct to simply pull the
correct id of the contextual model object instead
(which is what I did).

[#5773] fix teacher_wiki_and_tiny_images spec

This is an instance of using a string as an 
enumerable; adding a guard clause to
lib/sticky_sis_fields.rb to transform the string
argument into an array fixes it

[#5773] fix selenium/student_submissions_spec.rb

This is an issue with SSLCommon's use of 
net/https.  In older ruby versions there are 2 
parameters returned (the response, and the body).  
Now only the response is returned, 
so the response object has to be asked for the 
body explicitly. This fix was made in 
spec/selenium/helpers/files_common.rb

Change-Id: I66a25ebbc1d8e5af7b7117375210172d287a77c7
Reviewed-on: https://gerrit.instructure.com/14641
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>