properly handle time manipulations around DST-end
a265b7b reverted 37c15d2, which was the original version of this change.
This "reverts" a265b7b by restoring that functionality, but with
additional considerations.

First, the change is applied to readable date.js source, instead of the
minified version we had been using.

Second, this adds to the original commit to fix the set method on Date
objects to play nicely with the fixed addMilliseconds method.

In the vast majority of the year where both the old and the new
addMilliseconds behave identically, set worked fine. but on DST
boundaries where addMilliseconds was now fixed, it became incorrect for
set to use addMilliseconds/addSeconds/addMinutes/addHours instead of
using setMilliseconds/setSeconds/setMinutes/setHours.

In particular, when set tried to use addHours with a DST day as the
target, it would move e.g. 10 hours from midnight to either 9am or 11am
(depending on which DST boundary), rather than to 10am (which was either
9 or 11 hours away from midnight, depending on boundary).

This was particularly egregious since Date.parse ends up using today's
date as the base and then calling set on it. So all Date.parse calls --
that occurred on a DST-boundary day -- were broken.

fixes #11642

Change-Id: Ibc15796c699d1c3a1dea935f4ff68203270df484
Reviewed-on: https://gerrit.instructure.com/15025
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Clare Hetherington <clare@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>