add context to assignment cache key
this is important when it's a new assignment, because all new
assignments have the same cache key

test plan: visit on course assignments page to populate the cache, then
visit a different course assignments page and create a new assignment.
without refreshing, the assignment url and other data should be correct
for that course.

Change-Id: I457f2565d2855b9aed2950976ad4a3308e360707
Reviewed-on: https://gerrit.instructure.com/15029
Reviewed-by: Zach Pendleton <zachp@instructure.com>
Reviewed-by: Stanley Stuart <stanley+gerrit@instructure.com>
Tested-by: Brian Palmer <brianp@instructure.com>